backend/requirements.txt
APScheduler==3.11.2
bcrypt==4.0.1
fastapi==0.104.1
alembic==1.14.1
motor==3.3.2
SQLAlchemy==2.0.37
greenlet>=3.1.1
psycopg[binary]>=3.3.3
pymongo==4.6.1
passlib[bcrypt]==1.7.4
pydantic>=2.12.4
python-dotenv==1.0.0
email-validator>=2.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
qrcode[pil]>=7.4.2
resend>=2.0.0
boto3>=1.28.0
tzlocal==5.3.1
uvicorn==0.24.0
reportlab==4.4.10
pypdf
xlsxwriter==3.2.9
openpyxl==3.1.5
pandas==3.0.0
xlsxwriter==3.2.9
pywebpush
